extract logging to a Tracer type
Move tracing/log-level-watcher setup out of main.rs into a Tracer type
in logging.rs.
